feat: migrate DAAAccordionRow.spec
